pyproject.toml
[project]
name = "novel-rag"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.128.0",
    "llama-index>=0.14.12",
    "loguru>=0.7.3",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
    "zai-sdk>=0.2.0",
]
